+++ InterfaceGrafica/formCadastroProduto.cs
@@ -48,7 +48,7 @@ namespace FROGI_OS
             produtoRow.PRODUTO_CARACTERISTICA = pRODUTO_CARACTERISTICATextBox.Text;
             produtoRow.PRODUTO_UNIDADE_COMPRA = pRODUTO_UNIDADE_COMPRATextBox.Text;
             produtoRow.PRODUTO_UNIDADE_VENDA = pRODUTO_UNIDADE_VENDATextBox.Text;
-            try { produtoRow.PRODUTO_PESO = Convert.ToDecimal(pRODUTO_PESOTextBox.Text); } catch (Exception) { produtoRow.PRODUTO_PESO = 0; }
+            try { produtoRow.PRODUTO_PESO = Convert.ToDouble(pRODUTO_PESOTextBox.Text); } catch (Exception) { produtoRow.PRODUTO_PESO = 0; }
             produtoRow.PRODUTO_NUMERO = pRODUTO_NUMEROTextBox.Text;
             produtoRow.PRODUTO_EAN = pRODUTO_EANTextBox.Text;
             produtoRow.PRODUTO_NUMERO_SERIE = pRODUTO_NUMERO_SERIETextBox.Text;
@@ -60,9 +60,9 @@ namespace FROGI_OS
             try { produtoRow.PRODUTO_ESTOQUE_MINIMO = Convert.ToInt16(pRODUTO_ESTOQUE_MINIMOTextBox.Text); } catch (Exception) { produtoRow.PRODUTO_ESTOQUE_MINIMO = 0; }
             try { produtoRow.PRODUTO_ESTOQUE_MAXIMO = Convert.ToInt16(pRODUTO_ESTOQUE_MAXIMOTextBox.Text); } catch (Exception) { produtoRow.PRODUTO_ESTOQUE_MAXIMO = 0; }
             try { produtoRow.PRODUTO_ESTOQUE_ATUAL = Convert.ToInt16(pRODUTO_ESTOQUE_ATUALTextBox.Text); } catch (Exception) { produtoRow.PRODUTO_ESTOQUE_ATUAL = 0; }
-            try { produtoRow.PRODUTO_PRECO_COMPRA = Convert.ToDecimal(pRODUTO_PRECO_COMPRATextBox.Text); } catch (Exception) { produtoRow.PRODUTO_PRECO_COMPRA = 0; }
-            try { produtoRow.PRODUTO_PRECO_CUSTO = Convert.ToDecimal(pRODUTO_PRECO_CUSTOTextBox.Text); } catch (Exception) { produtoRow.PRODUTO_PRECO_CUSTO = 0; }
-            try { produtoRow.PRODUTO_PRECO_VENDA = Convert.ToDecimal(pRODUTO_PRECO_VENDATextBox.Text); } catch (Exception) { produtoRow.PRODUTO_PRECO_VENDA = 0; }
+            try { produtoRow.PRODUTO_PRECO_COMPRA = Convert.ToDouble(pRODUTO_PRECO_COMPRATextBox.Text); } catch (Exception) { produtoRow.PRODUTO_PRECO_COMPRA = 0; }
+            try { produtoRow.PRODUTO_PRECO_CUSTO = Convert.ToDouble(pRODUTO_PRECO_CUSTOTextBox.Text); } catch (Exception) { produtoRow.PRODUTO_PRECO_CUSTO = 0; }
+            try { produtoRow.PRODUTO_PRECO_VENDA = Convert.ToDouble(pRODUTO_PRECO_VENDATextBox.Text); } catch (Exception) { produtoRow.PRODUTO_PRECO_VENDA = 0; }
 
             if (dsFROGIOS.MARCA.Rows.Count != 0) produtoRow.PRODUTO_MARCA = ((dsFROGIOS.MARCARow)dsFROGIOS.MARCA.Rows[0]).MARCA_CODIGO;
             if (dsFROGIOS.GRUPO.Rows.Count != 0) produtoRow.PRODUTO_GRUPO = ((dsFROGIOS.GRUPORow)dsFROGIOS.GRUPO.Rows[0]).GRUPO_CODIGO;
